+++ CHANGELOG.md
@@ -23,6 +23,11 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ## [Unreleased]
 [Unreleased]: https://bitbucket.org/atlassian/aws-resources/branches/compare/master%0Drelease-1.3.1
 
+### Fixed
+- Stop detaching the policies from all stacks. Resolve [JPERF-317].
+
+[JPERF-317]: https://ecosystem.atlassian.net/browse/JPERF-317
+
 ## [1.3.1] - 2018-12-14
 [1.3.1]: https://bitbucket.org/atlassian/aws-resources/branches/compare/release-1.3.1%0Drelease-1.3.0